+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased](https://github.com/orisai/nette-monolog/compare/1.3.1...v1.x)
 
+### Changed
+
+- Tracy to PSR bridge - custom levels are mapped to info instead of error if the logged value is not an exception
+
 ## [1.3.1](https://github.com/orisai/nette-monolog/compare/1.3.0...1.3.1) - 2024-06-21
 
 ### Added